pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]

[tool.black]
line_length = 79

[tool.isort]
include_trailing_comma = true
line_length = 79
multi_line_output = 3

[tool.poetry]
authors = ["Luka Dijan <ldijan00@fesb.hr>", "Adriano Dijan <adijan00@fesb.hr>"]
description = ""
name = "lule"
version = "0.1.0"

[tool.poetry.dependencies]
aioboto3 = "^11.3.0"
envparse = "^0.2.0"
fastapi = "^0.66.0"
flake8 = "^3.9.2"
gino = {extras = ["pg", "starlette"], version = "^1.0.1"}
gunicorn = "^20.1.0"
mypy = "^0.9"
pre-commit = "^2.14.0"
pydantic = {extras = ["email"], version = "^1.8.2"}
python = "^3.10"
python-datauri = "^2.1.0"
uvicorn = "^0.14.0"

[tool.poetry.dev-dependencies]
alembic = "^1.6.5"
black = {version = "^21.9b0", allow-prereleases = true}
isort = "^5.9.3"
pytest = "^5.2"
requests = "^2.26.0"

[tool.poetry.group.dev.dependencies]
psycopg2 = "^2.9.7"
